+++ src/ggrc/models/deferred.py
@@ -1,0 +1,23 @@
+# Copyright (C) 2016 Google Inc.
+# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+
+"""Custom deferred wrapper.
+
+This deferred wrapper automatically adds the group name to a deferred object
+according to the current class.
+"""
+
+from ggrc import db
+
+
+def deferred(column, classname):
+  """Defer column loading for basic properties, such as boolean or string, so
+  that they are not loaded on joins. However, Identifiable provides an
+  eager_query implementation that will specify undefer in the options so that
+  when the resource is loaded itself, rather than through a join, it is
+  completely loaded.
+
+  In join tables, this function should not wrap foreign keys nor should it wrap
+  type properties for polymorphic relations.
+  """
+  return db.deferred(column, group=classname + '_complete')
